# BackendApplication/src/main/java/pl/edu/pwr/tkubik/ism/api/EventsController.java
package pl.edu.pwr.tkubik.ism.api;

import pl.edu.pwr.tkubik.ism.aspect.LogExecutionTime;
import pl.edu.pwr.tkubik.ism.aspect.LogMethod;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RestController;
import pl.edu.pwr.tkubik.ism.model.EventEntity;
import pl.edu.pwr.tkubik.ism.service.EventRegistrationService;
import pl.edu.pwr.tkubik.ism.service.EventService;
import pl.edu.pwr.tkubik.ism.service.EventCreationService;
import pl.edu.pwr.tkubik.ism.service.OrganizationService;
import pl.edu.pwr.tkubik.ism.repository.EventRegistrationRepository;
import pl.edu.pwr.tkubik.ism.api.EventApi;
import pl.edu.pwr.tkubik.ism.model.Event;

import java.time.LocalDate;
import java.time.LocalTime;
import java.util.ArrayList;
import java.util.List;

@RestController
public class EventsController implements EventApi {

    @Autowired
    private EventService eventService;

    @Autowired
    private EventRegistrationService eventRegistrationService;

    @Autowired
    private EventCreationService eventCreationService;

    @Autowired
    private OrganizationService organizationService;

    @Autowired
    private EventRegistrationRepository eventRegistrationRepository;

    private final Long CURRENT_USER_ID = 1L; // Utilisateur par défaut pour l'instant

    // private method to convert an Entity (JPA) to a DTO (OpenAPI)
    private Event convertToDto(EventEntity entity) {
        Event dto = new Event();
        dto.setEventId(entity.getId());
        dto.setTitle(entity.getTitle());

        // convert LocalDate (DB) to String (DTO)
        if (entity.getDate() != null) {
            dto.setDate(entity.getDate().toString());
        }

        // Convert LocalTime (DB) to String (DTO)
        if (entity.getTime() != null) {
            dto.setTime(entity.getTime().toString());
        }

        dto.setAddress(entity.getAddress());
        dto.setCategory(entity.getCategory());
        dto.setPrice(entity.getPrice());
        dto.setNumberOfPlaces(entity.getNumberOfPlaces());
        dto.setType(entity.getType());
        dto.setStatus(entity.getStatus());

        // On vérifie si l'utilisateur est inscrit et on l'ajoute au DTO
        boolean isRegistered = eventRegistrationRepository.existsByEventIdAndUserId(entity.getId(), CURRENT_USER_ID);
        // ATTENTION : Il faut t'assurer d'avoir ajouté le champ "isRegistered" et ses méthodes
        // get/set dans ta classe pl.edu.pwr.tkubik.ism.model.Event !
        // Si ce n'est pas encore fait, décommente la ligne ci-dessous une fois fait :
        dto.setIsRegistered(isRegistered);

        return dto;
    }

    // POST /events
    @LogMethod
    @LogExecutionTime
    @Override
    public ResponseEntity<Event> addEvent(Event eventDto) {
        // database Entity instance
        EventEntity entity = new EventEntity();

        // mappin DTO to Entity
        entity.setTitle(eventDto.getTitle());

        if (eventDto.getDate() != null) {
            entity.setDate(LocalDate.parse(eventDto.getDate()));
        }

        if (eventDto.getTime() != null) {
            entity.setTime(LocalTime.parse(eventDto.getTime()));
        }

        entity.setAddress(eventDto.getAddress());
        entity.setCategory(eventDto.getCategory());
        entity.setPrice(eventDto.getPrice());
        entity.setNumberOfPlaces(eventDto.getNumberOfPlaces());
        entity.setType(eventDto.getType());
        entity.setStatus(eventDto.getStatus());

        // call the service to save the event in the DB
        EventEntity savedEntity = eventService.addEvent(entity);

        // creating the link in the even_creations table

        // we suppose by default that organization with ID=1 create this event
        pl.edu.pwr.tkubik.ism.model.OrganizationEntity dummyOrg = new pl.edu.pwr.tkubik.ism.model.OrganizationEntity();
        dummyOrg.setId(1L);

        // creating the link
        pl.edu.pwr.tkubik.ism.model.EventCreation creation = new pl.edu.pwr.tkubik.ism.model.EventCreation();
        creation.setEvent(savedEntity);
        creation.setOrganization(dummyOrg);
        creation.setDateCreation(LocalDate.now());

        // save the link in DB
        eventCreationService.addCreation(creation);

        // retrieve the ID generated by the DB and set it in the DTO
        eventDto.setEventId(savedEntity.getId());

        // return the final DTO with the 200 HTTP status
        return new ResponseEntity<>(eventDto, HttpStatus.OK);
    }

    // POST /events/{eventId}/organizations/{orgId}
    @LogMethod
    @LogExecutionTime
    @Override
    public ResponseEntity<Event> addOrganizationToEvent(Long eventId, Long orgId) {

        // check if the event exists
        EventEntity event = eventService.findEventById(eventId);
        if (event == null) {
            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
        }

        // check if the organization exists
        pl.edu.pwr.tkubik.ism.model.OrganizationEntity organization = organizationService.findOrganizationById(orgId);
        if (organization == null) {
            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
        }

        // create the new link in the table "event_creations"
        pl.edu.pwr.tkubik.ism.model.EventCreation creation = new pl.edu.pwr.tkubik.ism.model.EventCreation();
        creation.setEvent(event);
        creation.setOrganization(organization);
        creation.setDateCreation(LocalDate.now());

        // save in DB
        eventCreationService.addCreation(creation);

        // return the event to confirm
        return new ResponseEntity<>(convertToDto(event), HttpStatus.OK);
    }

    // GET /events
    @LogMethod
    @LogExecutionTime
    @Override
    public ResponseEntity<List<Event>> findAllEvents(Double lat, Double lng, Integer radius, String category, String keyword, LocalDate date) {
        List<EventEntity> entities = eventService.findAllEvents();

        List<Event> dtos = new ArrayList<>();
        entities.forEach(entity -> dtos.add(convertToDto(entity)));

        return new ResponseEntity<>(dtos, HttpStatus.OK);
    }

    // GET /events/{eventId}
    @LogMethod
    @LogExecutionTime
    @Override
    public ResponseEntity<Event> findEventById(Long eventId) {
        EventEntity entity = eventService.findEventById(eventId);

        if (entity == null) {
            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
        }

        return new ResponseEntity<>(convertToDto(entity), HttpStatus.OK);
    }

    // PUT /events/{eventId}
    @LogMethod
    @LogExecutionTime
    @Override
    public ResponseEntity<Event> updateEvent(Long eventId, Event eventDto) {

        EventEntity existingEntity = eventService.findEventById(eventId);

        if (existingEntity == null) {
            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
        }

        existingEntity.setTitle(eventDto.getTitle());

        if (eventDto.getDate() != null) {
            existingEntity.setDate(LocalDate.parse(eventDto.getDate()));
        }
        if (eventDto.getTime() != null) {
            existingEntity.setTime(LocalTime.parse(eventDto.getTime()));
        }

        existingEntity.setAddress(eventDto.getAddress());
        existingEntity.setCategory(eventDto.getCategory());

        if (eventDto.getPrice() != null) {
            existingEntity.setPrice(eventDto.getPrice());
        }

        existingEntity.setNumberOfPlaces(eventDto.getNumberOfPlaces());
        existingEntity.setType(eventDto.getType());
        existingEntity.setStatus(eventDto.getStatus());

        EventEntity updatedEntity = eventService.addEvent(existingEntity);

        return new ResponseEntity<>(convertToDto(updatedEntity), HttpStatus.OK);
    }

    // DELETE /events/{eventId}
    @LogMethod
    @LogExecutionTime
    @Override
    public ResponseEntity<Event> deleteEventById(Long eventId) {
        EventEntity existingEntity = eventService.findEventById(eventId);

        if (existingEntity == null) {
            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
        }

        eventService.deleteEventById(eventId);

        return new ResponseEntity<>(convertToDto(existingEntity), HttpStatus.OK);
    }

    // POST /events/{eventId}/tickets
    @LogMethod
    @LogExecutionTime
    @Override
    public ResponseEntity<Event> registerForEvent(Long eventId, String body) {
        // check if the event exists
        EventEntity event = eventService.findEventById(eventId);

        if (event == null) {
            return new ResponseEntity<>(HttpStatus.NOT_FOUND); // Erreur 404
        }

        // we suppose by default that we use the user with ID=1
        pl.edu.pwr.tkubik.ism.model.UserEntity dummyUser = new pl.edu.pwr.tkubik.ism.model.UserEntity();
        dummyUser.setId(1L);

        // create the link
        pl.edu.pwr.tkubik.ism.model.EventRegistration registration = new pl.edu.pwr.tkubik.ism.model.EventRegistration();
        registration.setEvent(event);
        registration.setUser(dummyUser);
        registration.setDateRegistration(LocalDate.now());

        // save the link in the table "event_registrations"
        eventRegistrationService.addRegistration(registration);

        // we return the event to confirm and the HTTPO status ok
        return new ResponseEntity<>(convertToDto(event), HttpStatus.OK);
    }

    // PUT /tickets/{ticketId}/scan
    @LogMethod
    @LogExecutionTime
    @Override
    public ResponseEntity<Event> scanTicket(String ticketId, String body) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }
}