# BackendApplication/src/main/resources/openapi/openapi.yaml
openapi: 3.0.1
info:
  title: WrocłArts API
  description: REST API documentation for the WrocłArts system managing artistic activities in Wrocław.
  termsOfService: http://www.example.com/terms/
  contact:
    email: contact@maxime-luce.fr
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:8080/WroclArts
    description: Local Spring Boot Development Server
...
paths:
...
  /events:
    get:
      tags:
        - event
      summary: find all events (supports filtering and search)
      operationId: findAllEvents
      parameters:
        - name: lat
          in: query
          description: Latitude for distance filtering
          schema:
            type: number
            format: double
        - name: lng
          in: query
          description: Longitude for distance filtering
          schema:
            type: number
            format: double
        - name: radius
          in: query
          description: Search radius in meters
          schema:
            type: integer
        - name: category
          in: query
          description: Art category to filter
          schema:
            type: string
        - name: keyword
          in: query
          description: Keyword search
          schema:
            type: string
        - name: date
          in: query
          description: Date to filter events
          schema:
            type: string
            format: date
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Events'
            application/xml:
              schema:
                $ref: '#/components/schemas/Events'
      deprecated: false
      security:
        - wroclarts_auth:
            - read:events
    post:
      tags:
        - event
      summary: add a new event
      operationId: addEvent
      requestBody:
        description: Event object that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
          application/xml:
            schema:
              $ref: '#/components/schemas/Event'
        required: true
      responses:
        200:
          description: Event added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
            application/xml:
              schema:
                $ref: '#/components/schemas/Event'
        405:
          description: Validation exception
          content: {}
      security:
        - wroclarts_auth:
            - write:events
      x-codegen-request-body-name: event
  /events/{eventId}:
    get:
      tags:
        - event
      summary: find an event by eventId
      operationId: findEventById
      parameters:
        - name: eventId
          in: path
          description: Numeric ID of the event to get.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
            application/xml:
              schema:
                $ref: '#/components/schemas/Event'
        400:
          description: Invalid eventId supplied
          content: {}
        404:
          description: Event not found
          content: {}
      deprecated: false
      security:
        - wroclarts_auth:
            - read:events
    put:
      tags:
        - event
      summary: update an existing event
      operationId: updateEvent
      parameters:
        - name: eventId
          in: path
          description: Numeric ID of the event to update.
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Event object that needs to be updated
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
          application/xml:
            schema:
              $ref: '#/components/schemas/Event'
        required: true
      responses:
        200:
          description: Event updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
            application/xml:
              schema:
                $ref: '#/components/schemas/Event'
        404:
          description: Event not found
          content: { }
        405:
          description: Validation exception
          content: { }
      security:
        - wroclarts_auth:
            - write:events
      x-codegen-request-body-name: event

    delete:
      tags:
        - event
      summary: delete an event by eventId (Organization owner)
      operationId: deleteEventById
      parameters:
        - name: eventId
          in: path
          description: Numeric ID of the event to delete.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
            application/xml:
              schema:
                $ref: '#/components/schemas/Event'
        400:
          description: Invalid eventId supplied
          content: {}
        404:
          description: Event not found
          content: {}
      deprecated: false
      security:
        - wroclarts_auth:
            - write:events
  /events/{eventId}/organizations/{orgId}:
    post:
      tags:
        - event
      summary: Add an organization to an existing event (Co-host)
      operationId: addOrganizationToEvent
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Organization successfully added to the event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event or Organization not found
  /events/{eventId}/tickets:
    post:
      tags:
        - event
      summary: Register/Buy a ticket for an event
      operationId: registerForEvent
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Empty body for registration
        content:
          application/json:
            schema:
              type: string
          application/xml:
            schema:
              type: string
        required: false
      responses:
        200:
          description: Ticket generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
            application/xml:
              schema:
                $ref: '#/components/schemas/Event'
        400:
          description: Event fully booked or payment failed
          content: {}
      security:
        - wroclarts_auth:
            - write:events
      x-codegen-request-body-name: body
  /tickets/{ticketId}/scan:
    put:
      tags:
        - event
      summary: Scan a QR digital ticket to check-in
      operationId: scanTicket
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Empty body for scanning
        content:
          application/json:
            schema:
              type: string
          application/xml:
            schema:
              type: string
        required: false
      responses:
        200:
          description: Ticket successfully consumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
            application/xml:
              schema:
                $ref: '#/components/schemas/Event'
        400:
          description: Ticket already scanned or invalid
          content: {}
      security:
        - wroclarts_auth:
            - write:events
      x-codegen-request-body-name: body
...
  /organizations/{orgId}:
    get:
      tags:
        - organization
      summary: find an organization by orgId
      operationId: findOrganizationById
      parameters:
        - name: orgId
          in: path
          description: Numeric ID of the organization to get.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
            application/xml:
              schema:
                $ref: '#/components/schemas/Organization'
        400:
          description: Invalid orgId supplied
          content: { }
        404:
          description: Organization not found
          content: { }
      deprecated: false
      security:
        - wroclarts_auth:
            - read:organizations

    delete:
      tags:
        - organization
      summary: delete an organization by orgId
      operationId: deleteOrganizationById
      parameters:
        - name: orgId
          in: path
          description: Numeric ID of the organization to delete.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
            application/xml:
              schema:
                $ref: '#/components/schemas/Organization'
        400:
          description: Invalid orgId supplied
          content: { }
        404:
          description: Organization not found
          content: { }
      deprecated: false
      security:
        - wroclarts_auth:
            - write:organizations

    # --- LE NOUVEAU PUT EST ICI ---
    put:
      tags:
        - organization
      summary: update an existing organization profile
      operationId: updateOrganization
      parameters:
        - name: orgId
          in: path
          description: Numeric ID of the organization to update.
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Organization object to update
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Organization'
          application/xml:
            schema:
              $ref: '#/components/schemas/Organization'
        required: true
      responses:
        200:
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
            application/xml:
              schema:
                $ref: '#/components/schemas/Organization'
        404:
          description: Organization not found
          content: { }
        405:
          description: Validation exception
          content: { }
      security:
        - wroclarts_auth:
            - write:organizations
      x-codegen-request-body-name: organization
...
components:
  schemas:
...
    Event:
      required:
        - title
        - date
        - organizationsId
      type: object
      xml:
        name: Event
      properties:
        eventId:
          type: integer
          format: int64
        title:
          type: string
          example: Pottery Workshop
        date:
          type: string
          example: "2026-10-14"
        time:
          type: string
          example: "18:00"
        address:
          type: string
          example: Nadodrze
        category:
          type: string
          example: Crafts
        price:
          type: number
          format: double
          example: 120.0
        numberOfPlaces:
          type: integer
          example: 15
        type:
          type: string
          example: In-Person
        status:
          type: string
          example: Published
        isRegistered:
          type: boolean
          description: Is the current user registered
          example: false
    Events:
      type: array
      items:
        $ref: '#/components/schemas/Event'
      xml:
        name: Events
        wrapped: true
      example: |-
        <Events>
          <Event>
            <eventId>1</eventId>
            <title>Pottery Workshop</title>
            <date>2026-10-14</date>
            <time>18:00</time>
            <address>Nadodrze</address>
            <category>Crafts</category>
            <price>120.0</price>
            <numberOfPlaces>15</numberOfPlaces>
            <type>In-Person</type>
            <status>Published</status>
          </Event>
        </Events>
    Organization:
      required:
        - email
        - krsNumber
        - nameOrga
      type: object
      xml:
        name: Organization
      properties:
        orgId:
          type: integer
          format: int64
        email:
          type: string
          example: contact@ceramicstudio.pl
        nameOrga:
          type: string
          example: Ceramic Studio Wrocław
        krsNumber:
          type: string
          example: "0000123456"
        subscriptionDate:
          type: string
          format: date
          example: "2026-04-10"
        status:
          type: string
          example: PendingVerification
...
  securitySchemes:
    wroclarts_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://localhost/oauth/dialog
          scopes:
            write:events: modify events records in your account
            read:events: read your events records
            write:organizations: modify organization records
            read:organizations: read organization records
            write:products: modify shop products
            read:products: read shop products
            write:orders: modify order status
            read:orders: read order status
            admin:platform: exclusive platform administrator actions (moderation, verification)